requirements.txt
agno==1.6.1
aiofiles==24.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
backoff==2.2.1
bandit==1.8.3
beautifulsoup4==4.13.4
billiard==4.2.1
black==25.1.0
boto3==1.38.38
botocore==1.38.38
bs4==0.0.2
cachetools==6.0.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==45.0.4
dataclasses-json==0.6.7
decorator==5.2.1
distlib==0.3.9
distro==1.9.0
Django==4.2.21
django-cors-headers==4.7.0
django-debug-toolbar==4.3.0
django-filter==25.1
django-stubs==5.1.3
django-stubs-ext==5.1.3
django-unicorn==0.61.0
djangorestframework==3.16.0
docstring_parser==0.16
emoji==2.14.1
exceptiongroup==1.3.0
Faker==37.4.0
filelock==3.18.0
filetype==1.2.0
flake8==7.2.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
isort==6.0.1
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.0
kombu==5.5.3
langdetect==1.0.9
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.0
mypy==1.16.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
nodeenv==1.9.1
numpy==2.0.2
olefile==0.47
ollama==0.4.8
openai==1.78.0
orjson==3.10.18
packaging==25.0
pathspec==0.12.1
pbr==6.1.1
pgvector==0.4.1
pillow==11.2.1
platformdirs==4.3.8
pre_commit==4.2.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg==3.2.8
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pyOpenSSL==25.1.0
pypdf==5.5.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==6.0.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
s3transfer==0.13.0
service-identity==24.2.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.3
stevedore==5.4.1
textblob==0.19.0
tomli==2.2.1
tqdm==4.67.1
Twisted==25.5.0
txaio==23.1.1
typer==0.15.3
types-PyYAML==6.0.12.20250516
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.36.0
urllib3==1.26.20
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.2
zipp==3.21.0
zope.interface==7.2
